The name vars is too ambigous and hidden
I wanted to find a way to set an Environment variable for my Cloudflare worker via the wrangler cli / wrangler.toml and ended up looking at this block of example code.

This block of example code is too ambiguous for a simple task of setting an environment variable in the cloudflare. 

To create ONE basic environment variable,  I am presented with `kv_namespaces` and `[env.production]` and the ambiguous variable name `vars` does not help either.

Ideally the KV names spaces shouldn't be in this snippet, but cloudflare needs more users to adapt KV namespaces as it is a key part of their monetization strategy.

Hence, the least we could do is add a comment to make it more clear as to how to create environment variable.