Define TLD acronym before use (#552)
TLD is used to stand for top-level domain without the actual phrase being used anywhere. This change uses the phrase 'top-level domain' and indicates its abbreviaion to TLD before then using TLD going forward. This is a minor change that will help make this document and service more accessible to people with a non-technical background.