Update QR code image
This PR updates the image in the QR code tutorial to hide a reference to a previously-available domain. The demo link for this UI is available in the tutorial and the image is outdated, so this change will stop people from going to a URL that we aren't deploying to now.