Fix documented environment variable size limit
The environment variable size limit has always been, and remains, 1KB. At one time, we attempted to increase this to 5KB, but ran into some sort of issue, the details of which are unknown to me. Unfortunately, the 5KB figure crept into our docs. I tried to fix it once before, but I must have missed a couple places.

Also, changed KiB to KB for stylistic consistency with the rest of the document.