Remove confusing "limitation" about cross-object storage access
We don't ever intend to support objects reaching into each other's storage. It's confusing to list this as a "beta limitation" since that implies we intend to change it later. Also, multiple people have misinterpreted this as saying that objects cannot communicate directly with each other, and instead need to go back through a worker.