Fix to Python example since PyJWT 2.0.0
The Python example no long works since PyJWT 2.0.0. The code will run, but since we are catching all exceptions, devs will not notice that they must specify the `algorithms` parameter when decoding. The error will be hidden and hard to discover.

I have now included `algorithms` it in the example.