Update namecheap.md
There is no clickable link in the confirmation email - it simply says 

"If you WANT TO PROCEED with this transfer, you do not need to respond to this message. If you WISH TO CANCEL the transfer, please go to our website

If we do not hear from you by [date] at [time], the transfer will proceed.