Fix homebrew installation instructions
Per https://dev.to/gjrdiesel/getting-around-brew-s-error-calling-non-checksummed-download-of-17fl

If you follow the existing instructions, homebrew will complain with
`Calling Non-checksummed download of curl formula file from an arbitrary
URL is disabled!`.  The workaround (from the URL above) is to fetch the
file locally and then run `brew install --HEAD -s ...` on that.