Update deploy-a-jekyll-site.md
JEKYLL_ENV can be set in two way, as an environment variable or as part of the build command. Since the build command is generic to both the production and preview environments, an environment variable should be used instead.

JEKYLL_ENV is important for certain features to work properly in plugins and themes, and the documentation for Cloudflare Pages should specify how to set it as the default is for development environments and not production.