Fix codeblocks in ssh-cert-bastion.md
On the live site currently, a lot of the codeblocks have `sh`
as their language, which causes them to not be able to be
highlighted or copied.

There also are some which are actually shell commands but
still cannot be copied because they are missing the `$` at
the start of the block.

This PR fixes both of those issues.