Remove Non-Functioning Section
While this section does work when testing locally, it does not appear to work correctly when actually deploying with CloudFlare Pages.

After the initial `npm install` process is started inside CloudFlare any changes to the `.npmrc` file do not appear to be respected.

## Steps to Reproduce

I created the `.npmrc.pages` file following the guide here, and modified the `preinstall` script for the `package.json` according to the guide.

Here's the `.npmrc.pages` file that I used. `my_org` was actually the organization that I have private packages installed.
```
//npm.pkg.github.com/:_authToken=${TOKEN_FOR_GITHUB}
@my_org:registry=https://npm.pkg.github.com/my_org
```

I created and tested the token used for `TOKEN_FOR_GITHUB` and loaded it into CloudFlare Pages. 

I verified that running this locally correctly moved the `.npmrc.pages` file and installed packages.

I deployed to CloudFlare pages and received this error during the Building Application process
```
13:05:36.876	npm ERR! code E401
13:05:36.878	npm ERR! Unable to authenticate, need: Basic realm="GitHub Package Registry"
```

I manually copied over the  `.npmrc.pages` file to `.npmrc`, committed the results, and re-deployed. This time the deployment was successful. 

The conclusion here is that modifying the `.npmrc` in CloudFlare Pages using the `preinstall` is not a functional way to authenticate private registries, and CloudFlare Pages will need to make some changes on their side to support private registries. 

Using `//npm.pkg.github.com/:_authToken=${TOKEN_FOR_GITHUB}` in the `.npmrc` currently is not tenable due to the fact that it breaks all local development environments.