Update docs to aviod using "simple"
I think it's great to avoid using words like “easy”, “simple” and “basic” in docs, especially when Introducing a new method of a thing, because if users have a hard time completing the task that is supposedly “easy,” they will question their abilities. And we want people to feel comfortable learning.