Correct “Gigabytes per sec” to “Gigabyte-seconds”
Gb-s is appropriate for this, as the meaning is “the amount of memory used, multiplied by (or _over_) the total duration.” Thus in the example given, the memory used can be correctly expressed as *0.125 GB/s* (gigabytes per second) because the average amount of memory used in each second is 0.125GB. By contrast, what is intended is to communicate not the memory used per second, but instead the cumulative memory used, over the whole execution time, adding each second of time together. Gigabyte-seconds expresses this concept.