Changed example to restrict only resizing images from origin or host
As mentioned in https://chat.google.com/room/AAAAGX7FDAU/U0WF41wf64I, one of our customers mentioned that the code is an XSS vulnerability when straight copy & pasted. I don't agree that it is a full XSS exploit, but it would certainly make it easier to get things wrong with a straight copy and paste of the code.

The example will now only take an absolute pathname and will make the relative to either the origin if one is present and the hostname if not. This should make the example more secure by default.

I haven't been able to test this yet, so it would be great if someone could test it. I just wanted to get a PR in so we wouldn't forget.