Remove URL.createObjectURL from example script for reading POST data
Workers doesn't support URL.createObjectURL, so this example doesn't actually work on our platform.
Further, the example form doesn't actually accept a file upload, so this condition isn't useful on its own. 

This PR removes the line which uses that api, replacing it with a simple string return. Later, We might want a crack at rewriting this page altogether to be a more useful. example