Re-work the tutorial of exporting data from GraphQL API in CSV format (#2494)
This commit refactors a python tutorial and adds more details
a user should be aware of. It also removes not needed header and
query fields which makes the example simpler.

The initial python logic was not changed much and it still has
the same idea of querying last N day (with the specified offset)
of minutely aggregated attack activity.

Some texts were rephrased and indentation of the query was fixed
to make them clearer.