Update zero-trust-network-access.md
Updated the CIDR from an *unnatural* network boundary to a *natural* network boundary: 10.0.0.0/10 --> 10.0.0.0/8

Also, the 'excluded CIDR' (the last entry at the bottom) in the image:  ![Split Tunnel](../static/secure-web-gateway/split-tunnel/split-tunnel-entries.png) is really unnatural! "10.0.0.0/1". That exclusion would be: 0.0.0.0 --> 127.255.255.255. If that was the intended CIDR it would already encompass the top most exclusion from the same image. I believe the `split-tunnel` image should be updated to reflect 10.0.0.0/8 as seen in the following images: 
- ![Second network policy](../static/zero-trust-security/ztna/second-network-policy.png)
- ![Create network policy](../static/zero-trust-security/ztna/create-network-policy.png)