Updated "Caching Ports"
According to this: https://support.cloudflare.com/hc/articles/218411427

> Ports 80, 443, and 8080 are the only ports where Cloudflare Caching is available.

And this page itself, Port 8080 MUST be able to cache as it is listed as "HTTP ports supported by Cloudflare" and not in the "Caching is disabled for the following ports" section.

So should be pretty clear that the statement "Ports 80 and 443 are the only ports compatible with:" (Cloudflare Caching) cant be right, as they are not the only ones for "Cloudflare Caching". Please doublecheck and fix this as this can cause confusion.