Remove the reference to `worktop/cookie` in the cookie parsing example.
This doc gives a choice between 2 packages but isn't clear why one should be picked over the other (in fact, it claims they're the same). This confused a friend of mine who was looking at it, and I agreed it's confusing. This PR removes the reference to `worktop/cookie` in the example.