Clarification of remote and local machines
I followed the tutorial and felt that some clarification was needed between remote and local machines. As this was my first time carrying out such a task, at one point I began storing the public key to my local machine. It was not until I read on that I noticed the text mentioned that "While staying within the `/etc/ssh` directory". I then realised that the key needs to be stored on the remote machine and realised my error, and that it makes sense to do that. Therefore, I suggest some clarification just to make it clear what is done on the remote machine. Every other step was easy to follow. 

I have not added but it may be needed a line for logging onto the machine, e.g ssh vm.example.com. When I did this I got prompted for my password for my username on the machine. This may be something I had conflicting at one point, but when executing ssh user.name@vm.example.com where user.name matches the user preceding the domain from my SSO I get logged right in after SSO authorisation.