Updates - added note for new repos
Added these lines:

<Aside type="note">
If you have created a new repository, make sure you have created and pushed the `main` or `master` branch to it. The branch can be empty.
</Aside>

If user is using a new repository without any branches, user will not be able to proceed as "Production branch" dropdown would be empty.