Added note for frameworks
Added a note that states an issue with Cloudflare Pages according to which Redirects would only work if the site does not have any framework. 

The issue here is with the way some of these frameworks like Jekyll work, by generating a simple and static HTML/CSS/JS version of an entire project inside of a _site folder (in which we cannot write any extra file manually) and that's probably what is finally being deployed to Cloudflare's network. Because of this, even if the root folder of a project had a _redirects file, it would not be found in the deployed directory.