Update limits.md to reflect purge ratelimits
It's my understanding from talking to support and doing my own testing that purge-by-url doesn't count towards the 1200/5m limit. The docs were ambiguous whether the purge ratelimits are instead of, or in addition to, the global rate limit.

I think this is important information to spell out explicitly, since our team previously read this page to mean that the 1200/5m limit DID apply, which caused us to build unnecessary purging infrastructure, and might cause others to avoid using cloudflare.

Feel free to disregard this if the 1200/5m limit is meant to apply here, although be aware that it is not applied currently. Thanks!