[GraphQL API Docs] Improve tutorials
This commit updates api calls in tutorials so they can be used as
is from directly from a bash command line (accountTags and other
credentials should be provided though).

This commit also fixes a couple of tutorials that appeared to be
broken. All the queries were executed and successful results were
received.