Update _index.md
This has been a common confusion for customers, so it would help to have more clarity that URL Rewrite rules can't rewrite the host. 

I also reordered the page because I feel it makes more sense to have a brief description of Rewrite and Redirect at the top, and have the subsection describing static and dynamic rewrites.