Update WebRTC beta page to include invite to discord server
This was reported by a user in our Discord; they had trouble using
the link we provided originally, it only works if you were already
invited to the server.

This change tweaks that to provide a general invite to the server
(the same shared in our community docs) as well as the direct
channel link.