Simplify the Wrangler Getting Started page (#6157)
The previous page had a lot of information, but it wasn't necessarily geared towards someone getting started on a new Wrangler project or wanting to try out Wrangler as a tool. I've removed a lot of the unnecessary detail, and moved it into more appropriate sections (for instance, the details about logging in to Wrangler on a remote machine have been moved to the wrangler login section). Previously, it took a full page scroll to get to the first command—now, all the instructions for getting started are above the fold.