[Workers] Improve Request reconstruction example
The existing example dropped a few Request properties (that is, `fetcher`, `signal`, and `cf` -- plus any that we might add in the future), and threw an exception if the `event.request` object was a GET request with a body. Reconstructing a Request object by passing the original object as one of the two parameters to the Request constructor avoids both of these issues.