Fixes issue with home page SVG colors
On the developer doc home page the Queues product icon's coloring does not appear properly because the SVG fill is scoped only to paths. That particular icon uses rects resulting part of the icon not being filled with the correct color. This adds a simple fix to expand the styles to include rects.