[Docs] Fix issue in product list icon colors (#6420)
* Fixes issue with home page SVG colors

On the developer doc home page the Queues product icon's coloring does not appear properly because the SVG fill is scoped only to paths. That particular icon uses rects resulting part of the icon not being filled with the correct color. This adds a simple fix to expand the styles to include rects.

* Fix SVG rect fill color when using light theme

* Fix Reference Architecture icon (remove unnecessary rect)

Co-authored-by: Pedro Sousa <680496+pedrosousa@users.noreply.github.com>