[ZT] RDP port (#6634)
* Update rdp.md

The Connect as a User section here just isn't friendly with Windows users. 

It would seem that we're making an assumption here that we're not on a Windows machine, where I'd suspect the majority of user's are using Windows, if they're using RDP.  Also a majority of those user's likely also have port 3389 already open to RDP to that local machine.

This gets confusing.  Adding a note is the least we can do here.  I'd suggest that we rewrite this with a port thats other than 3389.

* style guide update

Co-authored-by: deansundquist <93232826+deansundquist@users.noreply.github.com>