Remove asides about missing wrangler / UI support for Queues
They're both now supported. Although there are still caveats --
the main one that comes to mind is support for queues in the preview
service. I don't think that merits mention in such a prominent place but
it would be something I'd document if we had a place we thought was good
for mentioning beta limitations. I guess the "Limits" page, but that's
more about limits on usage than things that don't work yet.