Update nameservers.md
Previously this guidance was necessary because of a catch-22 situation whereby they cannot set the nameservers at the registrar because the registrar detects the lack of a proper SOA response.

However, now this issue is no longer present - since we now respond with the Start of Authority (SOA) record on the correct set of nameservers for the pending domain. Therefore we should remove this section entirely.