[Pages] Clarify static asset request pricing in pricing.md
We had a small note about requests being free and unlimited, but it
wasn't as very clear and was hidden as the last sentence in the
Free Plan section. This adds a new Static Asset Requests header to the
Paid Plan to help clarify what a static request is, as well as where to
learn more about when Functions is used.