Add usage examples for Analytics Engine
As we grow Analytics Engine usage we're going to want to provide in-depth
examples of how customers can benefit. This commit adds the first of those
examples, which gives a basic summary of how and why a customer running a
software-as-a-service platform would use us. We should be able to use this
example as a direct reference with some customers.