Update wrangler docs for mtls file structure
While .pem files can contain multiple entities of different types
(certificates and private keys) due to the lack of a standard for the
format, Wrangler currently only accepts a .pem file with certificates
in it for `--cert` and a .pem file with a private key in it for `--key`
when using the `mtls-certificate upload` command.

This updates the documentation for the command to reflect that.

Also fixes a type in the definitions for the command

Resolves #8114