[Cache] Clarify meaning of Edge Cache TTL
Many customers believe that setting a high Edge Cache TTL of e.g. 1 year will make Cloudflare always cache their site content for a full year and raise queries to support and community when this is not the case.

My wording is a suggestion, I believe we should be clearer that the TTL is a maximum and not a guarantee. 

It may also be beneficial to direct users who need better guarantees to Cache Reserve, which will fulfil the 1 year TTL as long as there is 1 request every 30 days: https://developers.cloudflare.com/cache/about/cache-reserve/