Fix API docs schema behavior
This PR fixes an issue with the /schema function not correctly loading tags for specific endpoints. The previous functionality checked the first method for a given endpoint and tried to pull the tags out of it.

For instance, given `/users`, we would try and pull the tags from the `get` method defined underneath that `/users` endpoint.

This works until a team defines an endpoint and uses a ref to a component as the first "method" in an array of methods for an endpoint.

To fix this behavior, the function now iterates through each method in an endpoint, looking for tags. When one is found, it now updates the tag references correctly.

For reference, this is the identified endpoint that was susceptible to this bug, which should be viewable now in the deploy preview for this PR: `/api/operations/accounts-turnstile-widget-rotate-secret`.