Update move-domain.md
I'm seeing some tickets about customers that created accounts with a typo in their emails, added zones, and at some point, as some operations require a verified email they notice the typo, but then it is too late, and the only option we provide for this customers is to create a new account, and move the zones to the new account.

I think we should include in our documentation that the best way to move zones from one account to another account is by exporting DNS records before moving the zone, because some times, customers move the zones, and Cloudflare will import the proxied IPs instead of the origin IPs, which will generate a 1000 error. 

Here is a ticket example (https://cloudflare.zendesk.com/agent/tickets/2831953).