Document GraphQL API's new support for arrays of primitives
As widely requested, GraphQL API now supports arrays of primitive
types. This will be used for things like lists of tags.

This commit adjusts the existing documentation. It adds a list of
the new array operators: `_has`, `_hasall` and `_hasany`. It also
adjusts some documentation for Nested arrays, to try and disambiguate
between the limited filtering of those fields vs the more featured
filtering of the new arrays of primitives.