Update teams-troubleshooting.md
Chrome is rolling out an experimental feature that partitions the local storage in browsers. When this is enabled, the browser isolation client can send client storage to incorrect remote browsers, notably when tab switching. We're adding a new error dialog that will point users to this page to help them self-serve.