Update _product-limitations.md (#11678)
After a Live chat interaction with a PRE customer (ZD ticket 3016007), we realised the wording here could be improved to clarify that Version Management won't be available in case the zone is entitled to one of these features Waiting Room, China Network, WAF Attack Score or API Shield. This way being in sync with the current messaging on the dashboard as well. 
(PCX-8916)