Update protect-against-timing-attacks.md (#11617)
* Update protect-against-timing-attacks.md

Description:
The current documentation for the Crypto API in Cloudflare Workers suggests using crypto.timingSafeEqual(a, b) for performing a timing-safe comparison of two values. However, this method does not exist on the Crypto object. Instead, it is available through the SubtleCrypto interface.

This pull request updates the documentation to reflect the correct usage of timingSafeEqual. The proposed change includes an example and a brief explanation to guide developers on how to use this method correctly in the Cloudflare Workers environment.

Changes Made:

Updated the method call from crypto.timingSafeEqual(a, b) to crypto.subtle.timingSafeEqual(a, b)

* Update content/workers/examples/protect-against-timing-attacks.md

Co-authored-by: Kian <kian@kian.org.uk>

* Update protect-against-timing-attacks.md

implemented a change suggested by KianNH

* Update protect-against-timing-attacks.md

Co-authored-by: Kian <kian@kian.org.uk>

* Update content/workers/examples/protect-against-timing-attacks.md

Co-authored-by: Kian <kian@kian.org.uk>

* Update content/workers/examples/protect-against-timing-attacks.md

Co-authored-by: Kian <kian@kian.org.uk>

* Update protect-against-timing-attacks.md

---------

Co-authored-by: Kian <kian@kian.org.uk>