chore: Do not restrict search on product section
**🔖 Summary**

Following some recent improvements we made for the documentation's
search, this commit is a suggestion to remove an odd feature: currently,
when the search feature is triggered from a specific product section,
the search results which belong to that section are surfaced first, even
though they are less relevant.

From my experience, and according to others, this is rather confusing,
and this change is about preventing that behavior, effectively surfacing
the most relevant results, no matter which page the user is currently
browsing.

**🚧 Changes**

- Remove the Algolia [`optionalFilters`](https://www.algolia.com/doc/guides/managing-results/rules/merchandising-and-promoting/how-to/how-to-promote-with-optional-filters/)
  which were responsible for promoting lesser relevant search results
  from the currently browsed page.

**✅ Testing plan**

- Ran the documentation locally using `npm run dev` and looked at the
  before/after XHR requests going to Algolia to make sure the
  `optionalFilters` are not present anymore (also see the before/after
  screenshots showing what the search results look like).