change DO error handling docs to advise always retrying
the remote property isn't a reliable indicator for retrying.

Some information can be gleaned from it after doing a lot of parsing of
other DO-internal error messages that also set the remote property, but
this is ugly and too complex to document for users, since it's of little
benefit to them.

Co-authored-by: Maddy <130055405+Maddy-Cloudflare@users.noreply.github.com>