Switch back to `WebSocketHibernationServer` (#14497)
The class name needs to be the same across the script and
`wrangler.toml`. We could go with `WebSocketServer`, but then it looks
too similar to the regular websocket example. If a reader copied both
classes (regular and hibernatable) into a single script to test both
simultaneously, they would end up seeing an error because of naming
conflicts.