[Workers] Change wording around wrangler global flags
`wrangler` suports several global flags that supposedly
apply to all `wrangler` commands, such as `--version`
or `--help`. However, some of these flags are not
supported by Pages yet, and listing them as global is
causing confusion for our users.

This commit adjusts the wording around wrangler global
flags, to clarify which of these flags are not supported
by Pages.