Update the one-liner for Workers Analytics Engine
I was trying to understand the analytics options within Cloudflare (there are many!). I think the one-liner for Workers Analytics Engine today is confusing. 

Today, the one-liner is "Provides analytics about anything using Cloudflare Workers." This is confusing because it reads as follows: (Provides analytics about (anything using Cloudflare Workers)). Here is the issue with this one-liner:
* Workers Analytics Engine itself does not provide analytics. It is an engine to which events can be emitted, but does not provide analytics out of the box. 
* "anything using Cloudflare Workers" is confusing. What is using Cloudflare Workers? What is anything?