Update mobile-redirect.md
Hello team, per email sent to customers on April 19th:

"This message is to inform you that Cloudflare will deprecate and remove the Mobile Redirects product on June 15, 2024. To preserve the current functionality, you will need to migrate your Mobile Redirect configuration to Redirect Rules prior to June 15, 2024 or visitors will not be redirected to the preferred subdomain as per your existing Mobile Redirect configuration."

Thanks!