Update managed-networks.md
When you create a Managed Network Endpoint it is expected to be auto-excluded from all tunnels.  Meaning that if you attempt to otherwise access that particular endpoint over WARP -> Tunnel, it will not work. 

If the endpoint is shared or otherwise needs to be accessible over WARP -> Tunnel, a different endpoint should be chosen.