Update deploy-a-vitepress-site.mdx
In the original document, it is mentioned that `.vitepress` is added to the `.gitignore` file, which makes the entire `.vitepress` directory ignored and not uploaded to `github`.

However, the configuration file of `vitepress` is in this directory. Without the configuration file, the website built by `Cloudflare Pages` will not be able to correctly identify the route, and accessing the deployed website will only display a 404 page.

To ensure that the configuration file of `vitepress` is uploaded to `github` and other directories generated by local builds are ignored, just delete `.vitepress` from the `.gitignore` file and add `dist` and `cache`.