Adding steps to allow Warp connector to work as a private net offramp
Found during testing the current steps work for S2S connectivity, but if you try to access the private net behind the warp connector it fails due to traffic making it to the private net resources with a source IP in the 100.92.0.0/12 subnet, leading to the return route not being valid since the local subnet isn't aware of how to route to this subnet. Adding the command "sudo iptables -t nat -A POSTROUTING -o eth0 -j MASQUERADE" resolves this by changing the source IP of warp client traffic to the LAN IP of the warp connector, which private resources are able to locate.