Update minimum-tls.mdx
Adding examples to the minimum TLS for zone and hostname API endpoint calls.

We still see a lot of customers having trouble with the API syntax to set the minimum TLS version. The idea is to add a clear example to try to reduce the number of requests we get.

This way, we can simply point the customers to this document and they can copy and past the syntax on their end.