Fix Tor IP reputation documentation
Tor IP reputation is not solely based on the provided security level.
This commit updates the documentation to remove such specific call out,
which are part of Cloudflare internals.

If customers are looking to block Tor specifically, they should do so via a WAF
rule, which gives them fine grained control.