Remove Version minimum in update migrating-jekyll-from-github-pages.mdx
Cloudflare is running Ruby 3.2.2 which requires ~> 229 ( https://github.com/github/pages-gem/releases/tag/v229 )instead of 215 and was causing build errors.  

Seeing as this doc is really geared towards relative novices, I would argue for removing the version minimum in favor of latest.  That should be future proof for the majority of users who will be reading this doc.  More advanced users would be able to figure out the version pinning they need.