Extracts compatibility-flags out of compatibility-dates
While the concepts are related they are not the same.

compatibility-flags will grow as we had more flags and deserves a separate page. This also has the benefit to have add an entry in the side bar. Having to click on "Configuration > Compatibility dates" to get info on the Compatibility flags is counter-intuitive.

I have updated internal links to point to the newly added flag page. Unfortunately external links to a given flag (i.e. "workers/configuration/compatibility-dates/#nodejs-compatibility-flag") will not point to that flag but to the top of the "Compatibility date". This is a limitation of the redirect mechanism on the site which can not redirect a fragment.

There are only a few changes in the content for the new flag page. I extracted a sentence from the first section into as the subtitle and I fixed a typo. The only other change is updating the CODEOWNERS of the new page to the same as the page is was extracted from.