Cleaned up the Hostname priority
I wanted to make sure readers would not overlook this "Hostname priority" section, hence I removed (Cloudflare for SaaS) because it gives a false impression it's not needed to read it for a standard customer. In the end, it's very often the case that, somewhere, a customer have a hostname covered by Cloudflare for SaaS, but customer isn't aware of it.