Update cloudflare-gateway.mdx
The proposal is to also reference the private gateway default settings as bullet points, to make it clear which criteria is used and clearly know when it kicks in for private traffic. 
All the other sections (gateway to outbound internet and the overrides to gateway for private traffic) were already bullet points, so this consistency just helps customers parse better the criteria on the default behavior as well, instead of mistakenly looking at the possible overrides as the criteria.