Encourage changelogs to be written as part of docs PR template
Nudge people contributing to the developer docs to add changelogs.

- anything that's a new feature
- anything that's an important fix or improvement

We miss too many shots on goal today to share new things that have improved, too much good stuff happens and docs get updated but we don't share the news. Should encourage here as part of this template. Open to suggestions on framing.