Remove calls to sql.end() in Hyperdrive examples
As discussed, the workers runtime already handles closing outgoing
connections as soon as a worker invocation finishes. And we have reason
to believe that sql.end() can cause hangs due to a bug in postgres.js,
so it's better to just not include it.

I guess I didn't have to remove it from our node-postgres examples, but
it isn't needed there either, so why clutter our examples with extra
code?