Update cancel-subscription.mdx
Added adapted info from Macro: Billing > Moved domain does not cancel plan and account was invoiced

Furthermore, please note that changing the DNS on your domain name does not cancel your account, or your active subscription plans. The system waits for a manual cancellation of the subscriptions plans, since there are times that the owner of a domain name's DNS is required to take action to change it back because they did not intend to remove their account.  

You will receive a notification regarding the DNS change; however, that does not indicate the cancellation of your subscription plans.