Fix AutoRAG sub-folder filter
I tried the filter with `gt: "user/1//"`, however that didn't yield any files from "user/1/". When I exclude the second "/", it works as expected. I think the issue is that for files directly in the folder, the `gt` with trailing "/" doesn't catch, since the folder is exactly "user/1/", not "user/1/…".

Sample `response.data` from AutoRAG:

```json
  [
    {
      file_id: 'cb9b94f5-cc24-410f-97b0-8a1d7eb76613',
      filename: 'user/1/summer_clothing_line.csv',
      score: 0.642221,
      attributes: {
          timestamp: 1756299780000,
          folder: "users/1/",
          filename: "summer_clothing_line.csv"
      },
      content: […]
    }
  ]
```